Moves resource tests to testing package
Moves tests for Resource to the testing package. The Generate method was
moved to a new test file in the faststatus package to be exported for
tests.
